refactor: services (#1355)
