+++ src/locales/zh-Hant/messages.po
@@ -1232,4 +1232,3 @@ msgstr ""
 
 msgid "<0/> quoted your <1><2/></1>"
 msgstr ""
-
